Move pkg/appcatalog to pkg/input/appcatalog
