+++ pkg/input/appcatalog/appcatalog_test.go
