+++ pkg/input/appcatalog/testdata/giantswarm.yaml
