+++ pkg/input/appcatalog/types.go
